Changed Message field from sender id to sender login
Now the received Message from the database will have a sender field instead of sender_id.
Removed unnecessary default constructors and setters for models because they are not used anywhere.
User password and access token fields will not be retrieved anymore from the database for greater safety.
Also, removed access_token field from User because there is no need to save it in User now (but not in database!).